Added camera relative input & Freelook auto-orbit
The player always presses input relative to the camera's view, but as a
programmer it would be more practical to have that input relative to the
player character. There is now a MathUtilities static class with a
function to provide this functionality.

A freelook camera allows for the player to control the camera. However,
it is also common for the camera to slowly fall back behind the player
character as the player character moves around. While Unity's Freelook
provides this functionality out of the box, it's default implementation
takes too long for the camera to fall back behind the player.

I have added a FreelookRubberBand script which allows the camera to
auto-orbit behind the player faster.

If you want a fixed perspective that does not rotate with the player,
set the freelook to world space and don't use the FreelookRubberBand.

I have also created an example test scene that shows off the behavior.
The texture and shader came from the Unity ECS Physics samples. I have
not added any ECS to this project.

I modified the Input Manager so that the example can show off all the
features. The demo may break (easy fix) if you change the inputs, but
the actual non-demo code should still work.[Ma[b